+++ app/src/main/java/com/pechenkin/travelmoney/diagram/impl/DebitCreditDiagram.java
@@ -32,7 +32,7 @@ import java.util.List;
 public class DebitCreditDiagram extends Base {
     private boolean isAnimated = false;
 
-    public DebitCreditDiagram(double sum, List<Total.MemberSum> total) {
+    public DebitCreditDiagram(int sum, List<Total.MemberSum> total) {
         super(sum, total);
     }
