+++ app/src/main/java/com/pechenkin/travelmoney/diagram/impl/LineDiagram.java
@@ -36,7 +36,7 @@ public class LineDiagram extends Base {
 
     private boolean isAnimated = false;
 
-    public LineDiagram(double sum, List<Total.MemberSum> total) {
+    public LineDiagram(int sum, List<Total.MemberSum> total) {
         super(sum, total);
     }
